Refactor: lighthouse 측정 후 최적화를 위해 CarItem 컴포넌트 제거
